+++ inc/command/regeneratefilescommand.class.php
@@ -1,0 +1,53 @@
+<?php
+
+/**
+ * -------------------------------------------------------------------------
+ * Fields plugin for GLPI
+ * -------------------------------------------------------------------------
+ *
+ * LICENSE
+ *
+ * This file is part of Fields.
+ *
+ * Fields is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * Fields is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Fields. If not, see <http://www.gnu.org/licenses/>.
+ * -------------------------------------------------------------------------
+ * @copyright Copyright (C) 2013-2022 by Fields plugin team.
+ * @license   GPLv2 https://www.gnu.org/licenses/gpl-2.0.html
+ * @link      https://github.com/pluginsGLPI/fields
+ * -------------------------------------------------------------------------
+ */
+
+namespace PluginFieldsCommand;
+
+use Glpi\Console\AbstractCommand;
+use Symfony\Component\Console\Command\Command;
+
+class RegenerateFilesCommand extends AbstractCommand
+{
+    protected function configure()
+    {
+        $this->setName('plugin:fields:regenerate_files');
+        $this->setDescription(__('Regenerate container files.', 'fields'));
+        $this->setHelp(
+            __('Check all stored containers files (classes & front) are present and create them if needed', 'fields')
+            . "\n"
+        );
+    }
+
+    protected function execute($input, $output)
+    {
+        plugin_fields_checkFiles();
+        return Command::SUCCESS;
+    }
+}
